requirements.txt
# Core ML/AI frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.21.0

# CLIP and vision models
clip-by-openai>=1.0
sentence-transformers>=2.2.0

# Computer vision and image processing
opencv-python>=4.6.0
Pillow>=9.0.0
scikit-image>=0.20.0
imageio>=2.25.0

# 3D visualization and mesh processing
pyvista>=0.42.0
vtk>=9.2.0
trimesh>=3.20.0
meshio>=5.3.0

# Web framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Data science and utilities
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# Memory optimization (optional but recommended)
xformers>=0.0.20; sys_platform != "darwin"
bitsandbytes>=0.41.0; sys_platform == "linux"

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Additional utilities
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
